Various code changes to improve windows compatibility (#63)
* Fix some warnings in MSVC

* Fix some issues with our use of libcurl

* Fix error reporting
* Disable SSL peer validation in t_encrypt_compat; the vcpkg version of
  libcurl doesn't set up trust roots, and we don't really care too much
  for this use case.

* Disable mmap tests on Windows

While we can implement these later, they'll need to use a completely different
API (using e.g. VirtualProtect) instead.

* Remove variadic array usage

MSVC does not support C99 variadic arrays. Replace them with static
arrays and/or alloca.

* Fix t_encrypt_compat ignoring frame size

t_encrypt_compat was missing a call to set frame size, so it was only
testing the default frame size. This also turned up a bug in the test
case itself resulting in livelock for some framesizes due to an
incorrect condition controlling when the full message length was set on
the session.

* Minor CMakeLists fixes

* Fix AWS C++ SDK import
* Don't re-export openssl compiler/link flags when making a shared
library.
* Set up some warnings flags for MSVC

* Stop using hton* and switch to the aws-c-common byteswap routines

hton* have portability issues on windows; while they _can_ be used, they
require using the fragile WinSock2.h header (sensitive to include order,
among other things). This change switches to the aws-c-common byteswap
routines to sidestep these issues.

* Fix some more instances of restrict

+++ tests/lib/counting_mk.c
@@ -67,7 +67,7 @@ static int counting_mk_generate_data_key(
 
     enc_mat->unencrypted_data_key.len = aws_cryptosdk_alg_props(enc_mat->alg)->data_key_len;
     for (size_t i = 0; i < enc_mat->unencrypted_data_key.len; i++) {
-        enc_mat->unencrypted_data_key.buffer[i] = i;
+        enc_mat->unencrypted_data_key.buffer[i] = (uint8_t)i;
     }
 
     struct aws_cryptosdk_edk edk;
@@ -121,7 +121,7 @@ static int counting_mk_decrypt_data_key(struct aws_cryptosdk_mk * mk,
 
             dec_mat->unencrypted_data_key.len = props->data_key_len;
             for (size_t i = 0; i < dec_mat->unencrypted_data_key.len; i++) {
-                dec_mat->unencrypted_data_key.buffer[i] = i;
+                dec_mat->unencrypted_data_key.buffer[i] = (uint8_t)i;
             }
 
             return AWS_OP_SUCCESS;
